employee fetcher module
